# A faster implementation of NCC using cumulative summation