Runs 14x faster using a safer approach (to prevent overflow) and 24x faster using an unsafe approach (Tensorflow implementation). Also provide the option to use double precision (to prevent overflow).

Tested on Tensorflow (2.11.0) and PyTorch (2.1.0) with both GPU (RTX 3090 Ti) and CPU (i7-8700K). Attached are the run time and accuracy on GPU:

```
----------
Tensorflow
----------
Safe cumsum impl: 0.29789113998413086 s.
Unsafe cumsum impl: 0.17483901977539062 s.
Original impl: 4.185877799987793 s.
Safe cumsum vs original: [5.820766e-10].
Unsafe cumsum vs original: [5.820766e-10].
-------
PyTorch
-------
Safe cumsum impl: 0.10962605476379395 s.
Unsafe cumsum impl: 0.03331756591796875 s.
Original impl: 1.2561259269714355 s.
Safe cumsum vs original: 2.3283064365386963e-10.
Unsafe cumsum vs original: 0.0.
````